+++ datafusion/core/src/physical_optimizer/sort_pushdown.rs
@@ -40,7 +40,7 @@ use datafusion_physical_expr::{
 /// This is a "data class" we use within the [`EnforceSorting`] rule to push
 /// down [`SortExec`] in the plan. In some cases, we can reduce the total
 /// computational cost by pushing down `SortExec`s through some executors. The
-/// object carries the parent required ordering and the (optional) fetch value
+/// object carries the parent required ordering and the (optional) `fetch` value
 /// of the parent node as its data.
 ///
 /// [`EnforceSorting`]: crate::physical_optimizer::enforce_sorting::EnforceSorting
