+++ datafusion/physical-plan/src/sorts/sort.rs
@@ -26,7 +26,6 @@ use std::sync::Arc;
 
 use crate::common::spawn_buffered;
 use crate::expressions::PhysicalSortExpr;
-use crate::limit::statistics_with_fetch;
 use crate::metrics::{
     BaselineMetrics, Count, ExecutionPlanMetricsSet, MetricBuilder, MetricsSet,
 };
@@ -922,7 +921,7 @@ impl ExecutionPlan for SortExec {
     }
 
     fn statistics(&self) -> Result<Statistics> {
-        statistics_with_fetch(
+        Statistics::with_fetch(
             self.input.statistics()?,
             self.schema(),
             self.fetch,
